refactor: split KubernetesResourceUtil + added coverage
