+++ jkube-kit/config/resource/src/main/java/org/eclipse/jkube/kit/config/resource/ResourceVersioning.java
@@ -20,9 +20,12 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
 /**
  * @author nicola
- * @since 24/05/2017
  */
 @Builder(setterPrefix = "with")
 @AllArgsConstructor
@@ -31,6 +34,22 @@ import lombok.ToString;
 @Setter
 @ToString
 public class ResourceVersioning {
+
+  private static final Map<String, Function<ResourceVersioning, String>> KIND_API_VERSION_MAPPINGS = new HashMap<>();
+  static {
+    KIND_API_VERSION_MAPPINGS.put("Ingress", ResourceVersioning::getExtensionsVersion);
+    KIND_API_VERSION_MAPPINGS.put("StatefulSet", ResourceVersioning::getAppsVersion);
+    KIND_API_VERSION_MAPPINGS.put("Deployment", ResourceVersioning::getAppsVersion);
+    KIND_API_VERSION_MAPPINGS.put("NetworkPolicy", ResourceVersioning::getNetworkingVersion);
+    KIND_API_VERSION_MAPPINGS.put("Job", ResourceVersioning::getJobVersion);
+    KIND_API_VERSION_MAPPINGS.put("DeploymentConfig", ResourceVersioning::getOpenshiftV1version);
+    KIND_API_VERSION_MAPPINGS.put("CronJob", ResourceVersioning::getCronJobVersion);
+    KIND_API_VERSION_MAPPINGS.put("CustomResourceDefinition", ResourceVersioning::getApiExtensionsVersion);
+    KIND_API_VERSION_MAPPINGS.put("ClusterRole", ResourceVersioning::getRbacVersion);
+    KIND_API_VERSION_MAPPINGS.put("ClusterRoleBinding", ResourceVersioning::getRbacVersion);
+    KIND_API_VERSION_MAPPINGS.put("Role", ResourceVersioning::getRbacVersion);
+    KIND_API_VERSION_MAPPINGS.put("RoleBinding", ResourceVersioning::getRbacVersion);
+  }
 
   private String coreVersion;
   private String extensionsVersion;
@@ -42,5 +61,8 @@ public class ResourceVersioning {
   private String cronJobVersion;
   private String apiExtensionsVersion;
 
+  public String getForKind(String kind) {
+    return KIND_API_VERSION_MAPPINGS.getOrDefault(kind, ResourceVersioning::getCoreVersion).apply(this);
+  }
 }
