+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/complete-directory/pong-rc.yaml
