+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/contains_kind.yml
