+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/contains_no_kind.yml
