+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/invalid-metadata-pod.yaml
@@ -1,0 +1,1 @@
+metadata: ['invalid', 'not', 'an', 'array']
