+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/named-svc.yaml
