+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/network-v1-ingress.yml
