+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/networking-v1-np.yml
