+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/override-kind-pod.yml
@@ -1,0 +1,4 @@
+apiVersion: custom.jkube.eclipse.org/v1
+kind: NotAPod
+metadata:
+  name: not-a-pod
