+++ jkube-kit/enricher/api/src/test/resources/kubernetes-resource-fragments/simple-rc.yaml
