+++ jkube-kit/enricher/generic/src/test/java/org/eclipse/jkube/enricher/generic/DependencyEnricherTest.java
@@ -22,6 +22,7 @@ import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.resource.PlatformMode;
 import org.eclipse.jkube.kit.enricher.api.JKubeEnricherContext;
 import org.eclipse.jkube.kit.enricher.api.model.KindAndName;
+import org.eclipse.jkube.kit.enricher.api.util.KubernetesResourceFragments;
 import org.eclipse.jkube.kit.enricher.api.util.KubernetesResourceUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -86,8 +87,7 @@ class DependencyEnricherTest {
          * Our override file also contains a ConfigMap item with name jenkins, load it while
          * loading Kubernetes resources.
          */
-        return KubernetesResourceUtil.readResourceFragmentsFrom(
-                PlatformMode.kubernetes,
+        return KubernetesResourceFragments.readResourceFragmentsFrom(
                 KubernetesResourceUtil.DEFAULT_RESOURCE_VERSIONING,
                 "the-project",
                 resourceList.toArray(new File[0]));
