+++ jkube-kit/resource/helm/src/main/java/org/eclipse/jkube/kit/resource/helm/HelmService.java
@@ -41,7 +41,7 @@ import org.eclipse.jkube.kit.common.util.KubernetesHelper;
 import org.eclipse.jkube.kit.common.util.ResourceUtil;
 import org.eclipse.jkube.kit.common.util.Serialization;
 import org.eclipse.jkube.kit.config.resource.ResourceServiceConfig;
-import org.eclipse.jkube.kit.enricher.api.util.KubernetesResourceUtil;
+import org.eclipse.jkube.kit.enricher.api.util.KubernetesResourceFragments;
 
 import com.google.common.collect.Streams;
 import io.fabric8.kubernetes.api.model.HasMetadata;
@@ -234,7 +234,7 @@ public class HelmService {
 
   private static void splitAndSaveTemplate(Template template, File templatesDir) throws IOException {
     for (HasMetadata object : Optional.ofNullable(template.getObjects()).orElse(Collections.emptyList())) {
-      String name = KubernetesResourceUtil.getNameWithSuffix(KubernetesHelper.getName(object),
+      String name = KubernetesResourceFragments.getNameWithSuffix(KubernetesHelper.getName(object),
           KubernetesHelper.getKind(object)) + YAML_EXTENSION;
       File outFile = new File(templatesDir, name);
       ResourceUtil.save(outFile, object);
