+++ quickstarts/maven/spring-boot-crd/pom.xml
@@ -26,7 +26,7 @@
 
   <groupId>org.eclipse.jkube.quickstarts.maven</groupId>
   <artifactId>spring-boot-crd</artifactId>
-  <version>1.13.1</version>
+  <version>1.14-SNAPSHOT</version>
   <name>Eclipse JKube :: Quickstarts :: Maven :: Spring Boot - Custom Resources</name>
   <packaging>jar</packaging>
