+++ quickstarts/maven/spring-boot-crd/src/main/jkube/quarkus.yml
