+++ quickstarts/maven/spring-boot-crd/src/main/jkube/spring-boot.yml
