+++ quickstarts/maven/spring-boot-crd/src/main/jkube/vertx.yml
